Creating Curricula

Creating Curricula

CurricularAnalytics.add_requisite! — Function.

add_requisite!(rc, tc, requisite_type)

Add course rc as a requisite, of type requisite_type, for target course tc.

Requisite types

One of the following requisite types must be specified for rc:

pre : a prerequisite course that must be passed before tc can be attempted.

co : a co-requisite course that may be taken before or at the same time as tc.

strict_co : a strict co-requisite course that must be taken at the same time as tc.

source

add_requisite!([rc1, rc2, ...], tc, [requisite_type1, requisite_type2, ...])

Add a collection of requisites to target course tc.

Requisite types

The following requisite types may be specified for rc:

pre : a prerequisite course that must be passed before tc can be attempted.

co : a co-requisite course that may be taken before or at the same time as tc.

strict_co : a strict co-requisite course that must be taken at the same time as tc.

source

CurricularAnalytics.delete_requisite! — Function.

delete_requisite!(rc, tc)

Remove course rc as a requisite for target course tc. If rc is not an existing requisite for tc, an error is thrown.

Requisite types

The following requisite types may be specified for rc:

pre : a prerequisite course that must be passed before tc can be attempted.

co : a co-requisite course that may be taken before or at the same time as tc.

strict_co : a strict co-requisite course that must be taken at the same time as tc.

source

CurricularAnalytics.create_graph! — Function.

create_graph!(c::Curriculum)

Create a curriculum directed graph from a curriculum specification. The graph is stored as a LightGraph.jl implemenation within the Curriculum data object.

source

CurricularAnalytics.isvalid_curriculum — Function.

isvalid_curriculum(c::Curriculum, errors::IOBuffer)

Tests whether or not the curriculum graph associated with curriculum c is valid. Returns a boolean value, with true indicating the curriculum is valid, and false indicating it is not.

If c is not valid, the reason(s) why are written to the errors buffer. To view these reasons, use:

julia> errors = IOBuffer()
julia> isvalid_curriculum(c, errors)
julia> println(String(take!(errors)))

There are two reasons why a curriculum graph might not be valid:

Cycles : If a curriculum graph contains a directed cycle, it is not possible to complete the curriculum.

Extraneous Requisites : These are redundant requisites that introduce spurious complexity. If a curriculum has the prerequisite relationships $c_1 \rightarrow c_2 \rightarrow c_3$ and $c_1 \rightarrow c_3$, and $c_1$ and $c_2$ are not co-requisites, then $c_1 \rightarrow c_3$ is redundant and therefore extraneous.

source