CurricularAnalytics Data Types

CurricularAnalytics Data Types

CurricularAnalytics.Course — Type.

The Course data type is used to represent a single course consisting of a given number of credit hours. To instantiate a Course use:

Course(name, credit_hours; <keyword arguments>)

Arguments

name::AbstractString : the name of the course.

credit_hours::int : the number of credit hours associated with the course.

prefix::AbstractString : the prefix associated with the course.

num::AbstractString : the number associated with the course.

institution:AbstractString : the name of the institution offering the course.

canonical_name::AbstractString : the common name used for the course.

Examples:

julia> Course("Calculus with Applications", 4; prefix="MA", num="112", canonical_name="Calculus I")

source

CurricularAnalytics.Curriculum — Type.

The Curriculum data type is used to represent the collection of courses that must be be completed in order to earn a particualr degree. To instantiate a Curriculum use:

Curriculum(name, courses; <keyword arguments>)

Arguments

name::AbstractString : the name of the curriculum.

courses::Array{Course} : the collection of required courses that comprise the curriculum.

degree_type::Degree : the type of degree, allowable types: AA, AS, AAS, BA, BS (default).

institution:AbstractString : the name of the institution offering the curriculum.

system_type::System : the type of system the institution uses, allowable types: semester (default), quarter.

CIP::AbstractString : the Classification of Instructional Programs (CIP) code for the curriculum. See: https://nces.ed.gov/ipeds/cipcode

Examples:

julia> Curriculum("Biology", courses; institution="South Harmon Tech", degree_type=AS, CIP="26.0101")

source

CurricularAnalytics.LearningOutcome — Type.

The LearningOutcome data type is used to associate a set of learning outcomes with a course or a curriculum. To instantiate a LearningOutcome use:

LearningOutcome(name, description, hours)

Arguments

name::AbstractString : the name of the learning outcome.

description::AbstractString : detailed description of the learning outcome.

hours::int : number of class (contact) hours needed to attain the learning outcome.

Examples:

julia> LearningOutcome("M1", "Learner will demonstrate the ability to ...", 12)

source

CurricularAnalytics.Term — Type.

The Term data type is used to represent a single term within a DegreePlan. To instantiate a Term use:

Term([c1, c2, ...])

where c1, c2, ... are Course data objects

source

CurricularAnalytics.DegreePlan — Type.

The DegreePlan data type is used to represent the collection of courses that must be be completed in order to earn a particualr degree. To instantiate a Curriculum use:

DegreePlan(name, curriculum, terms, additional_courses)

Arguments

name::AbstractString : the name of the degree plan.

curriculum::Curriculum : the curriculum the degree plan must satisfy.

terms::Array{Term} : the arrangement of terms associated with the degree plan.

additional_courses::Array{Course} : additional courses in the degree plan that are not a part of the curriculum. E.g., a prerequisite math class to the first required math class in the curriculum.

Examples:

julia> DegreePlan("Biology 4-year Degree Plan", curriculum, terms)

source