Visualizing Curricula & Degree Plans

Visualizing Curricula and Degree Plans

CurricularAnalytics.visualize — Function.

visualize(degree_plan; <keyword arguments>))

Function used to visualize degree plans.

Arguments

Required:

degree_plan::DegreePlan : the name of the curriculum.

Keyword:

window : funtion argument that specifies the window to render content in. Default is Window().

changed : callback function argument, called whenever the curriculum is modified through

the interface. Default is nothing.

file_name : name of the file, in JSON format, that will the degree plan, including modifications. Default is recent-visualization.json.

source